Arreglado lo de los cast de Int a Long. Además se añade un método para eliminar todos los permisos a un rol. Esto se usa para actualizar los permisos, y que sean igual a lo que muestra el árbol cuando el usuario guarda los cambios.

+++ src/java/com/xedrux/cclouds/web/controllers/EmpresaController.java
@@ -86,7 +86,7 @@ public class EmpresaController {
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public HashMap<String, Object> updateEmpresa(
-            @ModelAttribute("Id") int id,
+            @ModelAttribute("Id") long id,
             @RequestBody @Valid CcloudsEmpresa empresa,
             BindingResult result)
             throws EntityNotFoundException, UnableToCreateEntityException {
